requirements.txt
Pillow>=10.0.0
icnsutil>=1.1.0
vtracer>=0.6.0